changed prec int to float in info data class model

+++ app/src/main/java/com/example/weatherapplication/model/Info.kt
@@ -3,7 +3,7 @@ package com.example.weatherapplication.model
 data class Info(
     val climate: String,
     val hum: Int,
-    val prec: Int,
+    val prec: Float,
     val temp: Int,
     val wind: Int
 )
